Moved logger to separate module
