+++ test/executable/executableTests.ts
@@ -20,7 +20,8 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 import * as semver from "semver";
-import { Logger, Options, run, Status } from "../../src/runner";
+import { Logger } from "../../src/logger";
+import { Options, run, Status } from "../../src/runner";
 import { denormalizeWinPath } from "../../src/utils";
 import { getNormalizedTypescriptVersion } from "../../src/verify/parse";
 import { createTempFile } from "../utils";
@@ -603,7 +604,9 @@ describe("Executable", function(this: Mocha.ISuiteCallbackContext) {
 
         if (semver.satisfies(tsVersion, ">=2.9")) {
             it("does not try to parse JSON files with --resolveJsonModule with TS >= 2.9", async () => {
-                const status = await execRunner({project: "test/files/tsconfig-resolve-json-module/tsconfig.json"});
+                const status = await execRunner({
+                    project: "test/files/tsconfig-resolve-json-module/tsconfig.json",
+                });
                 assert.equal(status, Status.Ok, "process should exit without an error");
             });
         }
